requirements.txt
Django==4.1.5
Django==4.2.3
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
python-dotenv==1.0.0